Implement `--version` arg
When git is available, the build system tries to get the information to
generate the version string from `git describe --tags`. The format of
the version taken from git is `x.x.x-dev.n+h`, where `n` is the commit
number (starting with the last tag) and where `h` is the hash of the
last commit.

If git is unavailable or an error occurs, the version string defaults
to the one written in `build.zig`.